python_backend/pyproject.toml
[project]
name = "content-creator-backend"
version = "0.1.0"
description = "Content Creator FastAPI Backend"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.29.0",
    "cloudinary>=1.36.0",
    "elevenlabs>=2.27.0",
    "fastapi>=0.115.0",
    "google-genai>=1.56.0",
    "httpx>=0.27.0",
    "langchain>=1.2.0",
    "langchain-anthropic>=1.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=1.1.0",
    "langchain-deepseek>=1.0.0",
    "langchain-google-genai>=4.0.0",
    "langchain-google-vertexai>=2.0.0",
    "langchain-groq>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.0",
    "langgraph-checkpoint-postgres>=3.0.0",
    "openai>=2.14.0",
    "openpyxl>=3.1.0",
    "pillow>=10.0.0",
    "psycopg[binary]>=3.2.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "python-multipart>=0.0.9",
    "supabase>=2.10.0",
    "tweepy>=4.14.0",
    "uvicorn[standard]>=0.30.0",
    "imageio-ffmpeg>=0.6.0",
    "facebook-business>=24.0.1",
    "google-adk>=1.0.0",
    "pywin32>=311; sys_platform == 'win32'",
    "tavily-python>=0.3.0",
    "pyyaml>=6.0.0",
    "deepagents>=0.3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
